requirements.txt
tqdm
notebook
openai
qdrant-client>=1.14.2
fastembed
python-dotenv
ipywidgets
pandas
scikit-learn
requests
streamlit
#sentence-transformers
#seaborn